+++ CHANGELOG.md
@@ -20,8 +20,11 @@ Development version
 - Support for `FEniCS 2019.1.0`
 - Modify `PointwiseObservation` so that ordering of observations targets is respected also in parallel.
   Setting the flag `prune_and_sort` to `True` restores previous behavior.
+- Remove unused input `tol` from `Model.solveFwd`, `Model.solveAdj`, `Model.solveFwdIncremental`, `Model.solveAdjIncremental`
+  and from related classes.
 - Use `argparse` to set parameters in application drivers from command line
 - Use `dl.XDMFFile` to export solutions for visualization in Paraview in all application drivers
+
 
 Version 2.3.0, released on Sept 6, 2019
 ----------------------------------------